Changed named of the project, fixed bug encountered by spectrum
