--- src/main/java/com/codeup/capstonestarter/CapstoneStarterApplication.java
@@ -1,13 +1,0 @@
-package com.codeup.capstonestarter;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CapstoneStarterApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CapstoneStarterApplication.class, args);
-    }
-
-}
