--- src/main/java/com/codeup/capstonestarter/data/errors/CustomAccessDeniedHandler.java
@@ -1,15 +1,0 @@
-package com.codeup.capstonestarter.data.errors;
-
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-public class CustomAccessDeniedHandler implements AccessDeniedHandler {
-
-    @Override
-    public void handle(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse, AccessDeniedException e) {
-        throw e;
-    }
-}
