--- src/main/java/com/codeup/capstonestarter/data/event/EventDates.java
@@ -1,42 +1,0 @@
-package com.codeup.capstonestarter.data.event;
-
-import com.codeup.capstonestarter.data.event.Event;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "event_dates")
-public class EventDates {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JsonIgnoreProperties("event_dates")
-    private Event event;
-
-    @Column(nullable = false)
-    private LocalDateTime date;
-
-    public EventDates() {
-    }
-
-    public Event getEvent() {
-        return event;
-    }
-
-    public void setEvent(Event event) {
-        this.event = event;
-    }
-
-    public LocalDateTime getDate() {
-        return date;
-    }
-
-    public void setDate(LocalDateTime date) {
-        this.date = date;
-    }
-}
