--- src/main/java/com/codeup/capstonestarter/data/post/Post.java
@@ -1,98 +1,0 @@
-package com.codeup.capstonestarter.data.post;
-
-import com.codeup.capstonestarter.data.comment.Comment;
-import com.codeup.capstonestarter.data.type.Type;
-import com.codeup.capstonestarter.data.user.User;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-import java.util.Collection;
-
-@Entity
-@Table(name = "posts")
-public class Post {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JsonIgnoreProperties({"events", "posts", "user"})
-    private Type type;
-
-    @Column(nullable = false)
-    private String title;
-
-    @Column(nullable = false, columnDefinition = "text")
-    private String body;
-
-    @Column(nullable = false)
-    private LocalDateTime createDate;
-
-    @ManyToOne
-    @JsonIgnoreProperties({"events", "email", "password", "bio", "postalCode", "comments", "posts", "types", "subEvents"})
-    private User user;
-
-    @OneToMany(mappedBy = "post", cascade = CascadeType.REMOVE)
-    private Collection<Comment> comments;
-
-    public Post(){}
-
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public Type getType() {
-        return type;
-    }
-
-    public void setType(Type type) {
-        this.type = type;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getBody() {
-        return body;
-    }
-
-    public void setBody(String body) {
-        this.body = body;
-    }
-
-    public LocalDateTime getCreateDate() {
-        return createDate;
-    }
-
-    public void setCreateDate(LocalDateTime createDate) {
-        this.createDate = createDate;
-    }
-
-    public User getUser() {
-        return user;
-    }
-
-    public void setUser(User user) {
-        this.user = user;
-    }
-
-    public Collection<Comment> getComments() {
-        return comments;
-    }
-
-    public void setComments(Collection<Comment> comments) {
-        this.comments = comments;
-    }
-}
