--- src/main/java/com/codeup/capstonestarter/data/user/User.java
@@ -1,197 +1,0 @@
-package com.codeup.capstonestarter.data.user;
-
-import com.codeup.capstonestarter.data.comment.Comment;
-import com.codeup.capstonestarter.data.event.Event;
-import com.codeup.capstonestarter.data.post.Post;
-import com.codeup.capstonestarter.data.type.Type;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-import javax.persistence.*;
-import javax.validation.constraints.Digits;
-import javax.validation.constraints.Email;
-import java.util.Collection;
-
-@Entity
-@Table(name = "users")
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String fullName;
-
-    @Column(nullable = false)
-    private String username;
-
-    @Email
-    @Column(nullable = false)
-    private String email;
-
-    @Column(nullable = false)
-    @JsonIgnore
-    private String password;
-
-    @Column(columnDefinition = "text")
-    private String bio;
-
-    @Column(nullable = false, length = 5)
-    private String postalCode;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private Role role = Role.u;
-
-    public enum Role {a, u}
-
-    @OneToMany(mappedBy = "user", cascade = CascadeType.REMOVE)
-    @JsonIgnoreProperties("user")
-    private Collection<Comment> comments;
-
-    @OneToMany(mappedBy = "user", cascade = CascadeType.REMOVE)
-    @JsonIgnoreProperties({"user"})
-    private Collection<Event> events;
-
-
-    @OneToMany(mappedBy = "user", cascade = CascadeType.REMOVE)
-    @JsonIgnoreProperties("user")
-    private Collection<Post> posts;
-
-    @ManyToMany(
-            fetch = FetchType.LAZY,
-            cascade = {CascadeType.MERGE, CascadeType.DETACH, CascadeType.PERSIST, CascadeType.REFRESH},
-            targetEntity = Type.class)
-    @JoinTable(
-            name="user_type",
-            joinColumns = {@JoinColumn(name = "user_id", nullable = false, updatable = false)},
-            inverseJoinColumns = {@JoinColumn(name="type_id", nullable = false, updatable = false)},
-            foreignKey = @ForeignKey(ConstraintMode.CONSTRAINT),
-            inverseForeignKey = @ForeignKey(ConstraintMode.CONSTRAINT)
-    )
-    @JsonIgnoreProperties("users")
-    private Collection<Type> types;
-
-    @ManyToMany(
-            fetch = FetchType.LAZY,
-            cascade = {CascadeType.MERGE, CascadeType.DETACH, CascadeType.PERSIST, CascadeType.REFRESH},
-            targetEntity = Event.class)
-    @JoinTable(
-            name="user_subEvents",
-            joinColumns = {@JoinColumn(name = "user_id", nullable = false, updatable = false)},
-            inverseJoinColumns = {@JoinColumn(name="sub_event_id", nullable = false, updatable = false)},
-            foreignKey = @ForeignKey(ConstraintMode.CONSTRAINT),
-            inverseForeignKey = @ForeignKey(ConstraintMode.CONSTRAINT)
-    )
-    @JsonIgnoreProperties("users")
-    private Collection<Event> subscribedEvents;
-
-    public User() {
-    }
-
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getFullName() {
-        return fullName;
-    }
-
-    public void setFullName(String fullName) {
-        this.fullName = fullName;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getBio() {
-        return bio;
-    }
-
-    public void setBio(String bio) {
-        this.bio = bio;
-    }
-
-    public String getPostalCode() {
-        return postalCode;
-    }
-
-    public void setPostalCode(String postalCode) {
-        this.postalCode = postalCode;
-    }
-
-    public Role getRole() {
-        return role;
-    }
-
-    public void setRole(Role role) {
-        this.role = role;
-    }
-
-    public Collection<Comment> getComments() {
-        return comments;
-    }
-
-    public void setComments(Collection<Comment> comments) {
-        this.comments = comments;
-    }
-
-    public Collection<Event> getEvents() {
-        return events;
-    }
-
-    public void setEvents(Collection<Event> events) {
-        this.events = events;
-    }
-
-    public Collection<Post> getPosts() {
-        return posts;
-    }
-
-    public void setPosts(Collection<Post> posts) {
-        this.posts = posts;
-    }
-
-    public Collection<Type> getTypes() {
-        return types;
-    }
-
-    public void setTypes(Collection<Type> types) {
-        this.types = types;
-    }
-
-    public Collection<Event> getSubEvents() {
-        return subscribedEvents;
-    }
-
-    public void setSubEvents(Collection<Event> subEvents) {
-        this.subscribedEvents = subEvents;
-    }
-}
