--- src/main/java/com/codeup/capstonestarter/data/user/UserRepository.java
@@ -1,12 +1,0 @@
-package com.codeup.capstonestarter.data.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    User findUserByUsername(String username);
-    Optional<User> findByEmail(String email);
-
-}
