+++ src/main/java/com/codeup/jecsnetwork/CapstoneStarterApplication.java
@@ -1,0 +1,13 @@
+package com.codeup.jecsnetwork;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CapstoneStarterApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CapstoneStarterApplication.class, args);
+    }
+
+}
