+++ src/main/java/com/codeup/jecsnetwork/data/errors/CustomAccessDeniedHandler.java
@@ -1,0 +1,15 @@
+package com.codeup.jecsnetwork.data.errors;
+
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.web.access.AccessDeniedHandler;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+public class CustomAccessDeniedHandler implements AccessDeniedHandler {
+
+    @Override
+    public void handle(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse, AccessDeniedException e) {
+        throw e;
+    }
+}
