+++ src/main/java/com/codeup/jecsnetwork/data/security/ResourceServerConfiguration.java
@@ -1,0 +1,47 @@
+package com.codeup.jecsnetwork.data.security;
+
+
+import com.codeup.jecsnetwork.data.errors.CustomAccessDeniedHandler;
+import com.codeup.jecsnetwork.data.errors.CustomAuthenticationEntryPoint;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer;
+import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfigurerAdapter;
+import org.springframework.security.oauth2.config.annotation.web.configurers.ResourceServerSecurityConfigurer;
+
+@Configuration
+@EnableResourceServer
+public class ResourceServerConfiguration extends ResourceServerConfigurerAdapter {
+
+    private final CustomAuthenticationEntryPoint customAuthenticationEntryPoint;
+
+    public ResourceServerConfiguration(CustomAuthenticationEntryPoint customAuthenticationEntryPoint) {
+        this.customAuthenticationEntryPoint = customAuthenticationEntryPoint;
+    }
+
+    @Override
+    public void configure(ResourceServerSecurityConfigurer resources) {
+        resources.resourceId("api");
+    }
+
+    @Override
+    public void configure(HttpSecurity http) throws Exception {
+        http
+                .formLogin()
+                .disable()
+                .sessionManagement()
+                .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+            .and()
+                .authorizeRequests()
+                 // .antMatchers("/api/users").hasAnyAuthority("ADMIN", "USER")
+                 // .antMatchers("/api/posts/**").hasAnyAuthority("ADMIN", "USER")
+                 // .antMatchers("/swagger-ui/**", "/v3/api-docs/**").permitAll()
+                 // .antMatchers("/api/users/create").permitAll()
+                .antMatchers("/**").permitAll()
+                .anyRequest().authenticated()
+            .and()
+                .exceptionHandling().authenticationEntryPoint(customAuthenticationEntryPoint).accessDeniedHandler(new CustomAccessDeniedHandler());
+    }
+
+}
