+++ src/main/java/com/codeup/jecsnetwork/data/security/UserService.java
@@ -1,0 +1,29 @@
+package com.codeup.jecsnetwork.data.security;
+
+import com.codeup.jecsnetwork.data.user.User;
+import com.codeup.jecsnetwork.data.user.UserRepository;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+import java.util.Arrays;
+
+@Service
+public class UserService implements UserDetailsService {
+
+    private final UserRepository repository;
+
+    public UserService(UserRepository repository) {
+        this.repository = repository;
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        User user = repository.findByEmail(email).orElseThrow(() -> new RuntimeException("User not found: " + email));
+        GrantedAuthority authority = new SimpleGrantedAuthority(user.getRole().name());
+        return new org.springframework.security.core.userdetails.User(user.getEmail(), user.getPassword(), Arrays.asList(authority));
+    }
+}
