+++ src/main/java/com/codeup/jecsnetwork/data/user/UserRepository.java
@@ -1,0 +1,12 @@
+package com.codeup.jecsnetwork.data.user;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    User findUserByUsername(String username);
+    Optional<User> findByEmail(String email);
+
+}
