+++ src/main/java/com/codeup/jecsnetwork/web/ViewController.java
@@ -1,0 +1,13 @@
+package com.codeup.jecsnetwork.web;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+public class ViewController {
+
+   @RequestMapping({"/", "/about", "/home", "/events", "/blog", "/account" })
+    public String showView() {
+       return "forward:/index.html";
+   }
+}
