--- src/test/java/com/codeup/capstonestarter/CapstoneStarterApplicationTests.java
@@ -1,13 +1,0 @@
-package com.codeup.capstonestarter;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CapstoneStarterApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
