+++ src/test/java/com/codeup/jecsnetwork/CapstoneStarterApplicationTests.java
@@ -1,0 +1,13 @@
+package com.codeup.jecsnetwork;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CapstoneStarterApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
